Fix buffer overrun in line input
Line array bumped up past 9 since scanf adds a null to the end. If we
increase it to 10 then it stops the buffer overrun but we can't tell if
a line is too long or not. So making it 100 solves that problem, and
gives enough context in the error message.